requirements.txt
pandas>=1.5.3
pytest>=7.2.2
sqlalchemy>=1.4.46
python-dotenv>=0.20.0
pymysql>=0.9.3